+++ src/mbapp.cpp
@@ -16,7 +16,7 @@ bool MicrobiomeApplication::run() {
     logger->log("Starting Microbiome Application");
 
     // run simulations
-    for (int i = 0; i < config.getNumSimulations(); i++) {
+    for (size_t i = 0; i < config.getNumSimulations(); i++) {
         std::cout << "Running simulation " << i + 1 << " of " << config.getNumSimulations() << std::endl;
         simulation = new Simulation(&config, i + 1, "Simulation " + std::to_string(i + 1));
         simulation->run();
@@ -41,7 +41,7 @@ bool MicrobiomeApplication::run() {
 
 void MicrobiomeApplication::printResults() {
     // print out result summaries
-    for (int i = 0; i < results.size(); i++) {
+    for (size_t i = 0; i < results.size(); i++) {
         std::cout << "=== Simulation " << i << " Result ===" << std::endl;
         std::cout << "Surviving Microorganisms: " << results[i].getSurvivingMicroorganisms() << std::endl;
         std::cout << "Dead Microorganisms: " << results[i].getDeadMicroorganisms() << std::endl;
@@ -56,7 +56,7 @@ void MicrobiomeApplication::printResultAverages() {
     int totalDeadMicroorganisms = 0;
     int totalEnergy = 0;
     int totalTicksElapsed = 0;
-    for (int i = 0; i < results.size(); i++) {
+    for (size_t i = 0; i < results.size(); i++) {
         totalSurvivingMicroorganisms += results[i].getSurvivingMicroorganisms();
         totalDeadMicroorganisms += results[i].getDeadMicroorganisms();
         totalEnergy += results[i].getEnergy();
